Exception handling fixes.
This catches a complex case: Suppose we have a trace
higher-order library function -> instrumented function -> library function
and the library function throws an exception that the instrumented
function does not handle. In this case, we must modify the stack
state of the higher-order function to ensure that it saves the exception
to be handled.