Vastly improve performance of points-to calculation.
The old algorithm would iterate over the whole version
map, which can grow quite big. This is incredibly slow
for longer traces.

Instead, we now walk along the object structure of
the new literals to figure out what needs to be
added.

Another improvement (not yet implemented) would be
to calculate the list of versioned references that
need to be visited in the version calculation
(since we know where changes occured there),
and just iterate over these lists here.